+++ src/main/java/com/mhamdi/brander/apis/dto/request/NewUserRequest.java
@@ -2,8 +2,6 @@ package com.mhamdi.brander.apis.dto.request;
 
 import java.util.List;
 
-import com.mhamdi.db.models.enums.AccountType;
-
 import lombok.Getter;
 import lombok.Setter;
 
@@ -14,5 +12,5 @@ public class NewUserRequest {
     private String username;
     private String password;
     // private AccountType role;
-    private List<Integer> templates;
+    private List<Long> templates;
 }
